re-organize committee/leader packages
move leader selection type to leader package
remove block translater, old committee type and unused functions
